+++ services/explorer-ui/src/pages/contract-class-details/json-viewer/json-display.tsx
@@ -1,0 +1,9 @@
+import { type FC } from "react";
+
+interface JsonDisplayProps {
+  data: unknown;
+}
+
+export const JsonDisplay: FC<JsonDisplayProps> = ({ data }) => {
+  return <pre className="overflow-auto">{JSON.stringify(data, null, 2)}</pre>;
+};
