+++ services/explorer-ui/src/pages/contract-class-details/key-value-helpers.ts
@@ -1,0 +1,43 @@
+import { ContractType, type ChicmozL2ContractClassRegisteredEvent } from "@chicmoz-pkg/types";
+import { routes } from "~/routes/__root";
+import { API_URL, aztecExplorer } from "~/service/constants";
+
+export const getContractClassKeyValueData = (
+  data: ChicmozL2ContractClassRegisteredEvent,
+) => {
+  return [
+    {
+      label: "BLOCK HASH",
+      value: data.blockHash,
+      link: `${routes.blocks.route}/${data.blockHash}`,
+    },
+    {
+      label: "CLASS ID",
+      value: data.contractClassId,
+    },
+    {
+      label: "VERSION",
+      value: data.version.toString(),
+    },
+    {
+      label: "ARTIFACT HASH",
+      value: data.artifactHash,
+    },
+    {
+      label: "PRIVATE FUNCTIONS ROOT",
+      value: data.privateFunctionsRoot,
+    },
+    {
+      label: "API ENDPOINT",
+      value: "View raw data",
+      extLink: `${API_URL}/${aztecExplorer.getL2ContractClassByIdAndVersion(
+        data.contractClassId,
+        data.version.toString(),
+      )}`,
+    },
+    {
+      label: "CONTRACT TYPE",
+      value: data.contractType ?? ContractType.Unknown,
+    },
+  ];
+};
