+++ services/explorer-ui/src/pages/contract-class-details/tabs-section.tsx
@@ -5,6 +5,7 @@ import { Loader } from "~/components/loader";
 import { OptionButtons } from "~/components/option-buttons";
 import { useDeployedContractInstances } from "~/hooks";
 import { contractClassTabs, type TabId } from "./constants";
+import { getDataFromMap } from "./display-utils";
 import { ContractInstancesTab } from "./tabs/contract-instances";
 import { ContractVersionsTab } from "./tabs/contract-versions";
 import { JsonTab } from "./tabs/json-tab";
@@ -24,19 +25,6 @@ interface TabSectionProps {
   isContractArtifactLoading: boolean;
   contractArtifactError: Error | null;
 }
-
-export const getDataFromMap = (data: SimplifiedViewOfFunc) => {
-  // eslint-disable-next-line @typescript-eslint/consistent-indexed-object-style
-  const result: { [key: string]: { [key: string]: string } } = {};
-  for (const [key, value] of data) {
-    // eslint-disable-next-line @typescript-eslint/consistent-indexed-object-style
-    result[key] = {};
-    for (const [k, v] of value) {
-      result[key][k] = v;
-    }
-  }
-  return result;
-};
 
 export const TabSection: FC<TabSectionProps> = ({
   contractClasses,
