+++ services/explorer-ui/src/pages/contract-class-details/tabs/json-tab.tsx
@@ -1,8 +1,10 @@
 import { type FC } from "react";
+import { JsonViewer } from "../json-viewer";
 
 interface JsonTabProps {
   data: unknown
 }
+
 export const JsonTab: FC<JsonTabProps> = ({ data }) => {
-  return <pre className="overflow-auto">{JSON.stringify(data, null, 2)}</pre>;
+  return <JsonViewer data={data} />;
 };
